# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: EffiPlat 全系统 API 文档
  version: "1.0"
  description: |
    EffiPlat 平台后端 API，涵盖用户、角色、权限、职责、环境、资产、服务、业务、Bug、认证等核心模块。
servers:
  - url: /api/v1
...
paths:
...
  /service-types:
    get:
      tags: [service-types]
      summary: 获取服务类型列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
          description: Page number for pagination.
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
          description: Number of items per page.
        - in: query
          name: name
          schema: { type: string }
          description: Filter by service type name (partial match).
        - in: query
          name: orderBy
          schema: { type: string, default: name, enum: [id, name, description, createdAt, updatedAt] }
          description: Field to sort by.
        - in: query
          name: sortOrder
          schema: { type: string, default: asc, enum: [asc, desc] }
          description: Sort order (ascending or descending).
      responses:
        '200':
          description: 成功获取服务类型列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTypeListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # 服务管理
  /services:
    get:
      tags: [services]
      summary: 获取服务列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
        - in: query
          name: name
          schema: { type: string }
        - in: query
          name: serviceTypeId
          schema: { type: integer }
        - in: query
          name: orderBy
          schema: { type: string, default: name, enum: [id, name, status, version, serviceTypeId, createdAt, updatedAt] }
          description: Field to sort by.
        - in: query
          name: sortOrder
          schema: { type: string, default: asc, enum: [asc, desc] }
          description: Sort order (ascending or descending).
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceListResponse'
    post:
      tags: [services]
      summary: 创建服务
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
  /services/{serviceId}:
    get:
      tags: [services]
      summary: 获取服务详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: serviceId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailsResponse'
    put:
      tags: [services]
      summary: 更新服务
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: serviceId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
    delete:
      tags: [services]
      summary: 删除服务
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: serviceId
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功，无响应体

  # 业务管理
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 错误的请求
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data: { type: [string, "null"] }
...
    Service:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: [string, "null"] }
        serviceType: { $ref: '#/components/schemas/ServiceType' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ServiceListResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Service' }
            total: { type: integer }
            page: { type: integer }
            pageSize: { type: integer }
...
    ServiceType:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ServiceTypeListResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ServiceType' }
            total: { type: integer }
            page: { type: integer }
            pageSize: { type: integer }

    # Asset Schemas
...
    ServiceCreateRequest:
      type: object
      required: [name, serviceTypeId]
      properties:
        name: { type: string }
        description: { type: [string, "null"] }
        serviceTypeId: { type: integer }
    ServiceUpdateRequest:
      type: object
      properties:
        name: { type: [string, "null"] }
        description: { type: [string, "null"] }
        serviceTypeId: { type: [integer, "null"] }
    ServiceResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          $ref: '#/components/schemas/Service'
    ServiceDetailsResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          $ref: '#/components/schemas/Service'
...